Fix URL segment

+++ app/Config/Routes.php
@@ -32,25 +32,25 @@ $routes->setAutoRoute(true);
 // We get a performance increase by specifying the default
 // route since we don't have to scan directories.
 // Auth
-$routes->get('/', 'Auth::index');
-$routes->get('logout', 'Auth::logout');
+$routes->get('/', 'Auth::index', ['filter' => 'is_auth']);
+$routes->get('logout', 'Auth::logout', ['filter' => 'not_auth']);
 
 // Home admin panel
-$routes->get('dashboard', 'Dashboard::index');
+$routes->get('dashboard', 'Dashboard::index', ['filter' => 'not_auth']);
 
 // User admin panel
-$routes->get('user', 'User::index');
-$routes->get('user/add', 'User::add');
-$routes->delete('user/(:num)', 'User::delete/$1');
-$routes->get('user/edit/(:segment)', 'User::edit/$1');
+$routes->get('user', 'User::index', ['filter' => 'not_auth']);
+$routes->get('user/add', 'User::add', ['filter' => 'not_auth']);
+$routes->delete('user/(:num)', 'User::delete/$1', ['filter' => 'not_auth']);
+$routes->get('user/edit/(:segment)', 'User::edit/$1', ['filter' => 'not_auth']);
 
 // Trash admin panel
-$routes->get('trash', 'Trash::index');
-$routes->get('trash/out', 'Trash::out');
-$routes->get('trash/add', 'Trash::add');
-$routes->delete('trash/(:num)', 'Trash::delete/$1');
-$routes->delete('trash/out/(:num)', 'Trash::delDataOut/$1');
-$routes->get('trash/edit/(:segment)', 'Trash::edit/$1');
+$routes->get('trash', 'Trash::index', ['filter' => 'not_auth']);
+$routes->get('trash/out', 'Trash::out', ['filter' => 'not_auth']);
+$routes->get('trash/add', 'Trash::add', ['filter' => 'not_auth']);
+$routes->delete('trash/(:num)', 'Trash::delete/$1', ['filter' => 'not_auth']);
+$routes->delete('trash/out/(:num)', 'Trash::delDataOut/$1', ['filter' => 'not_auth']);
+$routes->get('trash/edit/(:segment)', 'Trash::edit/$1', ['filter' => 'not_auth']);
 
 /**
  * --------------------------------------------------------------------
